add tests for label plots
