+++ composeml/tests/test_label_plots.py
@@ -1,10 +1,21 @@
-def test_distribution_plot(labels):
-    labels = labels.threshold(200)
-    plot = labels.plot.distribution()
-    assert plot.get_title() == 'Label Distribution'
+def test_count_by_time_categorical(total_spent):
+    labels = range(2)
+    total_spent = total_spent.bin(2, labels=labels)
+    ax = total_spent.plot.count_by_time()
+    assert ax.get_title() == 'Label Count vs. Cutoff Times'
 
 
-def test_count_by_time_plot(labels):
-    labels = labels.threshold(200)
-    plot = labels.plot.count_by_time()
-    assert plot.get_title() == 'Label Count vs. Time'
+def test_count_by_time_continuous(total_spent):
+    ax = total_spent.plot.count_by_time()
+    assert ax.get_title() == 'Label vs. Cutoff Times'
+
+
+def test_distribution_categorical(total_spent):
+    ax = total_spent.bin(2, labels=range(2))
+    ax = ax.plot.distribution()
+    assert ax.get_title() == 'Label Distribution'
+
+
+def test_distribution_continuous(total_spent):
+    ax = total_spent.plot.distribution()
+    assert ax.get_title() == 'Label Distribution'
